+++ src/test/java/com/microsoft/azure/functions/worker/functional/tests/SimpleParamReturnTest.java
@@ -2,31 +2,21 @@ package com.microsoft.azure.functions.worker.functional.tests;
 
 import com.microsoft.azure.functions.rpc.messages.*;
 import com.microsoft.azure.functions.worker.test.utilities.*;
-import org.junit.*;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
-import org.junit.runners.Parameterized.Parameters;
-import org.junit.runners.Parameterized.Parameter;
-import static org.junit.Assert.*;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
-@RunWith(Parameterized.class)
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 public class SimpleParamReturnTest extends FunctionsTestBase {
-
-	@Parameters
-	public static Object[] data() {
-		return new Object[] { "test String" };
-	}
-
-	@Parameter
-	public String stringInput;
 	private static String stringReturnValue;
 
 	public String ReturnStringFunction() {
 		return stringReturnValue;
 	}
 
-	@Test
-	public void testStringData() throws Exception {
+	@ParameterizedTest
+	@ValueSource(strings = {"test String"})
+	public void testStringData(String stringInput) throws Exception {
 		stringReturnValue = stringInput;
 		System.setProperty("azure.functions.worker.java.skip.testing", "true");
 		try (FunctionsTestHost host = new FunctionsTestHost()) {
