+++ src/test/java/com/microsoft/azure/functions/worker/test/ExecutionTraceContextTest.java
@@ -1,14 +1,14 @@
 package com.microsoft.azure.functions.worker.test;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.HashMap;
 import com.microsoft.azure.functions.TraceContext;
 import com.microsoft.azure.functions.worker.binding.ExecutionTraceContext;
+import org.junit.jupiter.api.Test;
 
-import org.junit.Test;
 
-public class ExecutionTraceContextTests {
+public class ExecutionTraceContextTest {
 
     @Test
     public void TraceContext_test_getAndset_nonEmpty() {
